requirements.txt
stmpy
paho-mqtt
tk